requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
sentence-transformers
chromadb
pydantic<2.0
python-dotenv
openai
google-generativeai
pypdf
